requirements.txt
aiohttp==3.11.18
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
discord-py-interactions==5.14.0
langchain-ollama==0.2.1
python-dotenv==1.0.1
langgraph==0.4.1
